requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.2
Django==5.0.6
django-ckeditor-5==0.2.12
django-cleanup==8.1.0
django-crispy-forms==2.1
django-environ==0.11.2
django-extensions==3.2.3
django-ninja==1.1.0
dnspython==2.6.1
email_validator==2.2.0
idna==3.7
iniconfig==2.0.0
iso3166==2.1.1
kombu==5.3.7
packaging==24.0
pillow==10.3.0
pluggy==1.4.0
prompt_toolkit==3.0.47
pydantic==2.7.1
pydantic_core==2.18.2
pydot==2.0.0
pyparsing==3.1.2
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
redis==5.0.7
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
